# src/main/resources/openapi/test-api.yaml
openapi: 3.0.1
info:
  title: PPAP Chat Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080 #기초 경로 springboot
...
paths:
...
  /api/users/signup:
    post:
      tags:
        - users
      summary: 사용자 가입
      description: 새로운 사용자를 가입합니다.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 가입 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusCode"
  /api/users/login:
    post:
      tags:
        - users
      summary: 사용자 로그인
      description: 사용자 로그인 처리
      operationId: userLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserToken'
        '400':
          description: 로그인 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusCode"
...
components:
  schemas:
...
    User:
      title: User Aggregate
      description: user info
      type: object
      #x-class-extra-annotation: "@jakarta.persistence.Entity"
      required: 
        - loginId
        - username
        - password
        - confirmPassword
        - email
        - userType
      properties:
        userId:
          type: integer
          format: int64
          description: user 테이블 pk
          #x-field-extra-annotation: "@jakarta.persistence.Id"
        loginId:
          type: string
          description: 사용자 아이디
        username:
          type: string
          description: 사용자 이름
        password:
          type: string
          description: 사용자 비밀번호
          #pattern: '^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[!@#\$%^&*])[a-zA-Z0-9!@#\$%^&*]{8,20}$'
          #example: "Passw0rd!"
        confirmPassword: 
          type: string
          description: 사용자 비밀번호 확인
        email:
          type: string
          description: 사용자 이메일
        isActive:
          type: boolean
          description: 사용 여부
          default: true
        isWithdrawed:
          type: boolean
          description: 탈퇴 여부
          default: false
        userType:
          type: string
          description: 개인 or 법인 사용자 타입
          enum:
            - PRIVATE
            - BUSINESS
          #x-constraints: "@ValidType"
        audit:
          $ref: '#/components/schemas/Audit'
    LoginDto:
      type: object
      description: 사용자 로그인 dto
      required:
        - loginId
        - password
      properties:
        loginId:
          type: string
        password:
          type: string
    UserToken:
      type: object
      description: 사용자 json web token
      required:
        - accessToken
        - refreshToken
        - grantType
        - userId
      properties:
        accessToken:
          type: string
          description: access token
        refreshToken:
          type: string
          description: refresh token
        grantType:
          type: string
          description: jwt 인증타입 bearer
        loginId:
          type: string
          description: 사용자 로그인 아이디
...
    Audit:
      type: object
      properties:
        createdDate:
          type: string
          format: date-time
          description: 생성일
        createdBy:
          type: string
          description: 생성자
        modifiedDate:
          type: string
          format: date-time
          description: 수정일
        modifiedBy:
          type: string
          description: 수정자
    StatusCode:
      type: object
      properties:
        '200':
          type: string
          example:
        '400':
          type: string
          example: Invalid user data provided
        '401':
          type: string
          example: Unauthorized access
        '500':
          type: string
          example: Internal server error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []